Split pixel world silhouette modules
